Interpolieren Funktion klappt

+++ Praktikum/app/src/main/java/com/example/praktikum/Interpolation.java
@@ -1,6 +1,7 @@
 package com.example.praktikum;
 
 import android.location.Location;
+import android.util.Log;
 
 import java.io.IOException;
 import java.sql.Array;
@@ -10,7 +11,7 @@ import java.util.List;
 
 public class Interpolation {
 
-    public List<Location> koordinatenLinearInterpolieren(Location a, Location b, int t1, int t2){
+    public List<Location> koordinatenLinearInterpolieren(Location a, Location b, long t1, long t2){
         List<Location> listeInterpolierteKoordinaten = new ArrayList<>();
         double dLongitude = b.getLongitude() - a.getLongitude();
         double dLatitude = b.getLatitude() - a.getLatitude();
@@ -20,7 +21,7 @@ public class Interpolation {
         long t = t1 + schrittweiteInMillisekunden;
 
         while(t < t2){
-            long dT = (t-t1)/t21;
+            double dT = (double)(t-t1)/(double) t21;
             double neueLongitude = a.getLongitude() + dLongitude * dT;
             double neueLatitude = a.getLatitude() + dLatitude * dT;
             Location neueKoordinate = new Location("Interpolation");
